+++ sepp-gui/src/main/java/de/verbund/sepp/gui/dialoge/DateiGUI.java
@@ -7,9 +7,6 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.IOException;
 import java.nio.file.attribute.FileTime;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
 
 import javax.swing.Icon;
 import javax.swing.JButton;
@@ -21,13 +18,12 @@ import de.verbund.sepp.main.daten.DateiInformationen;
 import de.verbund.sepp.main.daten.DatenSchnittstelle;
 import de.verbund.sepp.main.daten.DatenSchnittstelleImpl;
 import de.verbund.sepp.main.daten.Einstellungen;
+import de.verbund.sepp.main.utils.DatumHelfer;
 
 public class DateiGUI extends JFrame implements ActionListener {
 
 	Einstellungen einstellungen;
 	private JButton bInfo;
-	private static final DateTimeFormatter DATE_FORMATTER_WITH_TIME = DateTimeFormatter
-			.ofPattern("MMM d, yyyy HH:mm:ss");
 	private String verzeichnis = einstellungen.getProjektDateiPfad(); // TODO - Dateiaufruf
 	DatenSchnittstelle schnittstelle = DatenSchnittstelleImpl.getInstance();
 	DateiInformationen data = schnittstelle.getDateiInformationen(verzeichnis);
@@ -64,9 +60,9 @@ public class DateiGUI extends JFrame implements ActionListener {
 		JPanel p = new JPanel();
 		FileTime datum_1 = data.getErstellungsDatum();
 		FileTime datum_2 = data.getAenderungsDatum();
-		String s = fileTimeToString(datum_1);
+		String s = DatumHelfer.fileTimeToString(datum_1);
 		JLabel label_1 = new JLabel("Erstellungsdatum: " + s);
-		String s1 = fileTimeToString(datum_2);
+		String s1 = DatumHelfer.fileTimeToString(datum_2);
 		JLabel label_2 = new JLabel("Aenderungsdatum: " + s1);
 		p.add(label_1);
 		p.add(label_2);
@@ -80,25 +76,6 @@ public class DateiGUI extends JFrame implements ActionListener {
 		bInfo.addActionListener(this);
 		p.add(bInfo);
 		return p;
-	}
-
-	// DateTimeFormatter
-	public static String fileTimeToString(FileTime fileTime) {
-		String s = parseToString(fileTime.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime());
-		return s;
-	}
-
-	public static FileTime fileTimeFromString(String dateTimeString) {
-		LocalDateTime localDateTime = parseFromString(dateTimeString);
-		return FileTime.from(localDateTime.atZone(ZoneId.systemDefault()).toInstant());
-	}
-
-	public static String parseToString(LocalDateTime localDateTime) {
-		return localDateTime.format(DATE_FORMATTER_WITH_TIME);
-	}
-
-	public static LocalDateTime parseFromString(String date) {
-		return LocalDateTime.parse(date, DATE_FORMATTER_WITH_TIME);
 	}
 
 	@Override
