--- sepp-main/src/main/java/info.txt
@@ -1,1 +1,0 @@
-Damit das Verzeichnis gepushed wird
